# apps/cockpit/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IncidentAssignIn": {
        "properties": {
          "owner": {
            "title": "Owner",
            "type": "string"
          }
        },
        "required": [
          "owner"
        ],
        "title": "IncidentAssignIn",
        "type": "object"
      },
      "IncidentDetailOut": {
        "properties": {
          "check_name": {
            "default": "",
            "title": "Check Name",
            "type": "string"
          },
          "events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/IncidentEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opened_at": {
            "default": "",
            "title": "Opened At",
            "type": "string"
          },
          "owner": {
            "default": "",
            "title": "Owner",
            "type": "string"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "resolved_at": {
            "default": "",
            "title": "Resolved At",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "default": "fail",
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "default": "open",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "product"
        ],
        "title": "IncidentDetailOut",
        "type": "object"
      },
      "IncidentEventOut": {
        "properties": {
          "actor": {
            "default": "",
            "title": "Actor",
            "type": "string"
          },
          "at": {
            "default": "",
            "title": "At",
            "type": "string"
          },
          "detail": {
            "default": "",
            "title": "Detail",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind"
        ],
        "title": "IncidentEventOut",
        "type": "object"
      },
      "IncidentOut": {
        "properties": {
          "check_name": {
            "default": "",
            "title": "Check Name",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "opened_at": {
            "default": "",
            "title": "Opened At",
            "type": "string"
          },
          "owner": {
            "default": "",
            "title": "Owner",
            "type": "string"
          },
          "product": {
            "title": "Product",
            "type": "string"
          },
          "resolved_at": {
            "default": "",
            "title": "Resolved At",
            "type": "string"
          },
          "run_id": {
            "default": "",
            "title": "Run Id",
            "type": "string"
          },
          "severity": {
            "default": "fail",
            "title": "Severity",
            "type": "string"
          },
          "status": {
            "default": "open",
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "id",
          "product"
        ],
        "title": "IncidentOut",
        "type": "object"
      },
      "IncidentTransitionIn": {
        "properties": {
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "IncidentTransitionIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "DQ & Observability Cockpit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/incidents": {
      "get": {
        "description": "R4-1 Inbox: incidents from the lifecycle table, severity-sorted.",
        "operationId": "list_incidents_api_incidents_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IncidentOut"
                  },
                  "title": "Response List Incidents Api Incidents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Incidents",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/incidents/{incident_id}": {
      "get": {
        "operationId": "get_incident_api_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/incidents/{incident_id}/assign": {
      "post": {
        "description": "[AUTHZ] Assign an owner \u2014 steward role or higher.",
        "operationId": "assign_incident_api_incidents__incident_id__assign_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentAssignIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assign Incident",
        "tags": [
          "incidents"
        ]
      }
    },
    "/api/incidents/{incident_id}/transition": {
      "post": {
        "description": "[AUTHZ] Acknowledge/investigate/resolve \u2014 steward role or higher.",
        "operationId": "transition_incident_api_incidents__incident_id__transition_post",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-DQ-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Dq-Role"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentTransitionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition Incident",
        "tags": [
          "incidents"
        ]
      }
    },
...
  }
}